Mirroring
pm.polyMirrorFace
'direction' flag specifies the mirror direction.
'mergeMode' flag specifies the behaviour of the mirror with respect to the border edges. Valid values are 0-5, corresponding to +X, -X, +Y, -Y, +Z, -Z direction respectively. If the mode is 0, the border edges will not be merged (co-incident vertices will be present). If the mode is 1, the border vertices/edges will be merged If the mode is 2, the border edges will be extruded and connected.